Add files via upload
Analysis for the Beam Generator using a filled rate array for one to one weighting opposed to a singular multiplication factor. This is an old version of the analysis that should not be used but provides insight on how we can properly utilize weighting with boost histograms and awkward arrays.