Create EnergyBeamAnalysis.py
Analysis of the Energy distribution of particles with abs(hitpid)'s of 11, 13, and 211 in the BEAM generator. These are electrons and positrons, muons, and pions. 
NOTE: This script will take awhile to run and complete. TODO: Optimize it with uproot multithreading. If you need to run it, ensure the time requested is > 10:00:00 and you are running a minimum of 8 cpus.